[Refactoring] class 분리, gtest include 개선
- main.cpp추가 및 main 함수 이동
- EmployeeManagement의 class 정의/구현 분리
- CommandParser의 class 정의/구현 분리
- gtest include에 대해 pch.cpp와 pch.h에 정리
 : 공통 함수는 pch.cpp로 이동
 : pragama once 누락 header 수정
