+++ UnitTest/employeeManagement_test.cpp
@@ -1,5 +1,12 @@
 #include "pch.h"
-#include "../FriendsProject/employeeManagement.h"
+
+using namespace std;
+
+TEST(employeeManagementTest, test) {
+	EmployeeManagement* em = new EmployeeManagement();
+
+	em->testEM();
+}
 
 TEST(employeeManagementTest, open_file) {
 	EmployeeManagement* em = new EmployeeManagement();
