+++ UnitTest/pch.h
@@ -5,3 +5,11 @@
 #pragma once
 
 #include "gtest/gtest.h"
+#include "../FriendsProject/employeeinfo.h"
+#include "../FriendsProject/employeeManagement.h"
+#include "../FriendsProject/CommandParser.h"
+#include "../FriendsProject/dataManager.h"
+#include "../FriendsProject/search.h"
+#include "../FriendsProject/printer.h"
+
+void input_test_data(DataManager* data_manager);
