+++ UnitTest/search_test.cpp
@@ -1,5 +1,4 @@
 #include "pch.h"
-#include "../FriendsProject/search.cpp"
 
 EmployeeInfo employees[1000];
 SearchEngine searchEngine;
