--- UnitTest/test.cpp
@@ -1,7 +1,0 @@
-#include "pch.h"
-#include "../FriendsProject/employeeInfo.h"
-
-TEST(TestCaseName, TestName) {
-  EXPECT_EQ(1, 1);
-  EXPECT_TRUE(true);
-}
